fix(tasks): rename mention user folder
