+++ packages/sanity/src/tasks/src/tasks/components/form/FormEdit.tsx
@@ -19,7 +19,7 @@ import {useTasksNavigation} from '../../context'
 import {useRemoveTask} from '../../hooks/useRemoveTask'
 import {type TaskDocument} from '../../types'
 import {ActivityLog} from '../activityLog'
-import {AssigneeEditFormField} from './mentionUser/AssigneeEditFormField'
+import {AssigneeEditFormField} from './assignee'
 import {RemoveTaskDialog} from './RemoveTaskDialog'
 import {StatusSelector} from './StatusSelector'
 import {Title} from './TitleField'
